Rename hash and deserialize files
