+++ src/transaction/proto_deserialize.c
@@ -16,7 +16,7 @@
  *****************************************************************************/
 #include "buffer.h"
 
-#include "deserialize.h"
+#include "proto_deserialize.h"
 #include "utils.h"
 #include "types.h"
 #include "com/daml/ledger/api/v2/interactive/interactive_submission_service.pb.h"
@@ -30,11 +30,9 @@
 #include "ledger_assert.h"
 #endif
 
-parser_status_e transaction_deserialize(buffer_t *buf, transaction_t *tx) {
+parser_status_e proto_deserialize(buffer_t *buf, signing_type_e type, transaction_t *tx) {
     LEDGER_ASSERT(buf != NULL, "NULL buf");
     LEDGER_ASSERT(tx != NULL, "NULL tx");
-    const char memo[] =
-        "memo";  // Placeholder for memo, can be replaced with actual memo parsing logic
 
     if (buf->size > MAX_TX_LEN) {
         return WRONG_LENGTH_ERROR;
@@ -44,23 +42,33 @@ parser_status_e transaction_deserialize(buffer_t *buf, transaction_t *tx) {
 
     PRINTF("Decoding transaction from buffer of size %d bytes\n", buf->size);
 
+    const pb_msgdesc_t* message_type;
+
+    switch (type) {
+        case SIGN_PREPARED_TRANSACTION:
+            message_type = com_daml_ledger_api_v2_interactive_PrepareSubmissionResponse_fields;
+            break;
+        case SIGN_UNTYPED_VERSIONED_MESSAGE:
+            // TODO: implement when needed
+        case SIGN_HASH:
+        default:
+            return VALUE_PARSING_ERROR;
+    }
+
     if (!pb_decode(&stream,
-                   com_daml_ledger_api_v2_interactive_PrepareSubmissionResponse_fields,
+                   message_type,
                    &tx->prepared_tx)) {
         PRINTF("Failed to decode transaction: %s\n", PB_GET_ERROR(&stream));
         return VALUE_PARSING_ERROR;
     }
 
-    tx->memo_len = strlen(memo);
-    tx->memo = (uint8_t *) memo;  // Assigning a static memo for demonstration purposes
-
     PRINTF("Decoded transaction successfully.\n");
-    PRINTF("Transaction fields : \n");
-    PRINTF("Prepared TX hash: %.*H\n",
-           tx->prepared_tx.prepared_transaction_hash.size,
-           tx->prepared_tx.prepared_transaction_hash.bytes);
-    PRINTF("  Has prepared transaction: %d\n", tx->prepared_tx.has_prepared_transaction);
-    PRINTF("  Has hashing details: %d\n", tx->prepared_tx.hashing_details != NULL);
+    // PRINTF("Transaction fields : \n");
+    // PRINTF("Prepared TX hash: %.*H\n",
+    //        tx->prepared_tx.prepared_transaction_hash.size,
+    //        tx->prepared_tx.prepared_transaction_hash.bytes);
+    // PRINTF("  Has prepared transaction: %d\n", tx->prepared_tx.has_prepared_transaction);
+    // PRINTF("  Has hashing details: %d\n", tx->prepared_tx.hashing_details != NULL);
 
     return PARSING_OK;
 }
